+++ insertionSort.py
@@ -1,0 +1,11 @@
+# Type: Insertion Sort
+def insertion_sort(arr):
+    if len(arr) <= 1:
+        return
+    for i in range(1, len(arr)):
+        key = arr[i]
+        j = i - 1
+        while j >= 0 and key < arr[j]:
+            arr[j + 1] = arr[j]
+            j -= 1
+        arr[j + 1] = key
