+++ main.py
@@ -2,62 +2,11 @@ import random
 import timeit
 import matplotlib.pyplot as plt
 import numpy as np
+from hybridSort import hybridSort
 
-K_VALUES = [8, 16, 32, 64, 128, 256, 512, 1024]
-N_VALUES = [8_192, 16_384, 32_768, 65_536, 131_072, 262_144, 524_288, 1_048_576]
+K_VALUES = [20, 25, 30, 35, 40, 45, 50, 55, 60, 65, 70, 75, 80]
+N_VALUES = [500, 1_000, 5_000, 10_000, 50_000, 100_000, 500_000, 1_000_000]
 times_per_n = {n: [] for n in N_VALUES}
-
-# Merge Sort helper
-def merge(arr, l, m, r):
-    n1 = m - l + 1
-    n2 = r - m
-    L = arr[l:m + 1]
-    R = arr[m + 1:r + 1]
-
-    i, j, k = 0, 0, l  # Initial index of merged subarray
-
-    while i < n1 and j < n2:
-        if L[i] <= R[j]:
-            arr[k] = L[i]
-            i += 1
-        else:
-            arr[k] = R[j]
-            j += 1
-        k += 1
-
-    while i < n1:
-        arr[k] = L[i]
-        i += 1
-        k += 1
-
-    while j < n2:
-        arr[k] = R[j]
-        j += 1
-        k += 1
-
-# Insertion Sort
-def insertionSort(arr):
-    for i in range(1, len(arr)):
-        key = arr[i]
-        j = i - 1
-        while j >= 0 and key < arr[j]:
-            arr[j + 1] = arr[j]
-            j -= 1
-        arr[j + 1] = key
-
-def mergeSort(arr, l, r, k):
-    if l < r:
-        m = l + (r - l) // 2
-        hybridSort(arr, l, m, k)
-        hybridSort(arr, m + 1, r, k)
-        merge(arr, l, m, r)
-
-# Hybrid Sort
-def hybridSort(arr, l, r, k):
-    if len(arr) <= k:
-        insertionSort(arr)
-    elif l < r:
-        mergeSort(arr, l, r, k)
 
 def main():
     # Test code
@@ -70,7 +19,6 @@ def main():
                 time_to_sort = timeit.timeit(lambda: hybridSort(arr, 0, len(arr) - 1, k), number=1)
                 times.append(time_to_sort)
             avg_times_for_n.append(np.mean(times))
-        print(f"n={n}: {avg_times_for_n}")
         times_per_n[n] = avg_times_for_n
 
     # Plotting the performance of the hybrid sort
